% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMscale-package.R
\docType{data}
\name{biketrack}
\alias{biketrack}
\title{GPS recorded bike track}
\format{'data.frame':	254 obs. of  4 variables:\cr
$ lon : num  13 13 13 13 13 ...\cr
$ lat : num  52.4 52.4 52.4 52.4 52.4 ...\cr
$ time: POSIXct, format: "2016-05-18 07:53:22" "2016-05-18 07:53:23" ...\cr
$ ele : num  66 66 66 67 67 67 68 69 69 69 ....}
\source{
GPS track export from OSMtracker App
}
\description{
My daily bike route, recorded with the app OSMtracker on my Samsung Galaxy S5
}
\examples{

data(biketrack)
plot(biketrack[,1:2])
# see equidistPoints

}
\keyword{datasets}

