% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthDist.R
\name{earthDist}
\alias{earthDist}
\title{distance between lat-long coordinates}
\usage{
earthDist(df, r = 6371)
}
\arguments{
\item{df}{Dataframe with lat and long columns.
The distance of all the entries relative to the first row is computed.}

\item{r}{radius of the earth. Could be given in miles. DEFAULT: 6371 (km)}
}
\value{
Vector with distance(s) in km
}
\description{
Great-circle distance between points at lat-long coordinates.
(The shortest distance over the earth's surface).
}
\examples{
d <- read.table(header=TRUE, sep=",", text="
lat, long
52.514687, 13.350012   # Berlin
35.685024, 139.753365  # Tokio
51.503162, -0.131082") # London
earthDist(d) # 8922 and 928 km
map <- pointsMap(d, zoom=2)

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2016.
        Angle formula from Diercke Weltatlas 1996, Page 245
}
\seealso{
\code{\link{degree}} for pre-formatting,
         \url{http://www.movable-type.co.uk/scripts/latlong.html}
}
\keyword{spatial}

