% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSMscale-package.R
\docType{package}
\name{OSMscale-package}
\alias{OSMscale}
\alias{OSMscale-package}
\title{Add a Scalebar to OpenStreetMap Plots}
\description{
functionality to project points to UTM and add a scalebar to OpenStreetMap plots.
}
\note{
Get the most recent code updates at \url{https://github.com/brry}
}
\examples{

# see scaleBarOSM

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, June 2016
}
\keyword{documentation}
\keyword{package}

