% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleBar.R
\name{scaleBar}
\alias{scaleBar}
\title{scalebar for OSM plots}
\usage{
scaleBar(map, x = 0.1, y = 0.9, length = 0.2, abslen = NA,
  unit = c("km", "m", "mi", "ft", "yd"), field = "rect", fill = NA,
  adj = c(0.5, 1.5), cex = par("cex"), col = 1, targs = NULL, lwd = 5,
  lend = 1, ...)
}
\arguments{
\item{map}{Map object with map$tiles[[1]]$projection to get the projection from.}

\item{x, y}{Relative position of left end of scalebar. DEFAULT: 0.1, 0.9}

\item{length}{Approximate relative length of bar. DEFAULT: 0.2}

\item{abslen}{Absolute length in \code{unit}s. DEFAULT: NA (computed internally from length)}

\item{unit}{Unit for computation and label. kilometer and meter as well as
miles, feet and yards are possible.
Note that the returned absolute length is in m. DEFAULT: "km"}

\item{field, fill, adj, cex, col}{Arguments passed to \code{\link[berryFunctions]{textField}}}

\item{targs}{List of further arguments passed to \code{\link[berryFunctions]{textField}}
like font, col, etc. DEFAULT: NULL}

\item{lwd, lend}{Line width and end style passed to \code{\link{segments}}. DEFAULT: 5,1}

\item{\dots}{Further arguments passed to \code{\link{segments}} like col}
}
\value{
invisible: coordinates of scalebar and label
}
\description{
Add a scalebar to default or UTM-projected OpenStreetMap plots
}
\details{
This uses a hack to get the right distance in the default mercator projected maps.
There, the axes are not in meters, but rather ca 0.7m units (for NW Germany area maps with 20km across).
Accordingly, other packages plot wrong bars:\cr
SDMTools::Scalebar(x=1442638,y=6893871,distance=10000)\cr
raster::scalebar(d=5000, xy=c(1442638,6893871))\cr
mapmisc::scaleBar(map$tiles[[1]]$projection, seg.len=10, pos="center", bg="transparent")\cr
I suppose this function works for other projections as well, but haven't tried yet.
You might need to specify abslen manually with other projections where the axes do not resemble meters.
}
\examples{
d <- data.frame(long=c(12.95, 12.98, 13.22, 13.11), lat=c(52.40,52.52, 52.36, 52.45))
map <- pointsMap(d)
coord <- scaleBar(map)  ; coord
scaleBar(map, 0.3, 0.05, unit="m")
scaleBar(map, 0.3, 0.5, unit="km", length=0.1)
scaleBar(map, 0.12, 0.28, abslen=20000)
scaleBar(map, 0.6, 0.8, unit="mi", col="red", targ=list(col="blue"))

\dontrun{ ## Too much downloading time, too error-prone
# Tests around the world
par(mfrow=c(1,2), mar=rep(1,4))
long <- runif(1, -180, 180) ; long <- c(long,long+runif(1,0,4))
lat <- runif(1, -90, 90) ; lat <- c(lat,lat+runif(1,0,4))
# long <- c(5,15) ; lat=c(45,60)
map <- pointsMap(data.frame(long,lat))
map2 <- pointsMap(data.frame(long,lat), map=map, utm=TRUE)
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2016
}
\seealso{
\code{\link{pointsMap}}, \code{\link{projectPoints}}
}
\keyword{aplot}
\keyword{spatial}

