% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointsMap.R
\name{pointsMap}
\alias{pointsMap}
\title{Get map for lat-long points}
\usage{
pointsMap(data, x = "long", y = "lat", fx = 0.05, fy = fx,
  type = "osm", zoom = NULL, minNumTiles = 9L, mergeTiles = TRUE,
  map = NULL, utm = FALSE, proj = paste0("+proj=utm +zone=", zone,
  "+ellps=WGS84 +datum=WGS84"), zone = mean(long)\%/\%6 + 31, plot = TRUE,
  add = FALSE, scale = TRUE, quiet = FALSE, ...)
}
\arguments{
\item{data}{Data.frame with coordinates}

\item{x, y}{Names of columns in \code{data} containing longitude (East-West)
and latitude (North-South) coordinates. DEFAULT: "long","lat"}

\item{fx, fy}{Extend factors (additional map space around actual points)
passed to \code{\link{extendrange}}. DEFAULT: 0.05}

\item{type}{Tile server in \code{\link[OpenStreetMap]{openmap}}}

\item{zoom, minNumTiles, mergeTiles}{Arguments passed to \code{\link[OpenStreetMap]{openmap}}}

\item{map}{Optional map object. If given, it is not downloaded again. DEFAULT: NULL}

\item{utm}{Logical: Convert map to UTM (or other \code{proj})?
Consumes some extra time. DEFAULT: FALSE}

\item{proj}{proj4 character string or CRS object to project to.
Only used if utm=TRUE. DEFAULT: UTM projection at \code{zone}}

\item{zone}{UTM zone, see e.g. \url{https://upload.wikimedia.org/wikipedia/commons/e/ed/Utm-zones.jpg}.
DEFAULT: \link{mean} of \code{long}}

\item{plot}{Logical: Should map be plotted and points added? DEFAULT: TRUE}

\item{add}{Logical: add points to existing map? DEFAULT: FALSE}

\item{scale}{Logical: add \code{\link{scaleBar}}? DEFAULT: TRUE}

\item{quiet}{Logical: suppress progress messages? DEFAULT: FALSE}

\item{\dots}{Further arguments passed to \code{\link{points}} like pch, lwd, col, ...}
}
\value{
Map returned by \code{\link{openmap}}
}
\description{
Download and plot map with the extend of a dataset with lat-long coordinates
}
\examples{
d <- read.table(sep=",", header=TRUE, text=
"lat, long # could e.g. be copied from googleMaps, rightclick on What's here?
43.221028, -123.382998
43.215348, -123.353804
43.227785, -123.368694
43.232649, -123.355895")
map <- pointsMap(d)
scaleBar(map, ndiv=5)
map_utm <- pointsMap(d, map=map, utm=TRUE)
axis(1); axis(2) # now in meters
projectPoints(d$lat, d$long)
scaleBar(map_utm, x=0.2, y=0.8, unit="mi", type="line", col="red", length=0.25)
pointsMap(d[1:2,], map=map_utm, add=TRUE, col="red", pch=3, lwd=3)

d <- data.frame(long=c(12.95, 12.98, 13.22, 13.11), lat=c(52.40,52.52, 52.36, 52.45))
map <- pointsMap(d, type="bing") # aerial map

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2016
}
\seealso{
\code{\link{projectPoints}}, \code{\link[OpenStreetMap]{openmap}}
}
\keyword{hplot}
\keyword{spatial}

