% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.R
\name{proj}
\alias{pll}
\alias{posm}
\alias{proj}
\alias{putm}
\title{CRS of various PROJ.4 projections}
\usage{
putm(long, zone = mean(long, na.rm = TRUE)\%/\%6 + 31)

posm()

pll()
}
\arguments{
\item{long}{Vector of decimal longitude coordinates (East/West values).
Not needed of \code{zone} is given.}

\item{zone}{UTM (Universal Transverse Mercator) zone, see e.g. \url{https://upload.wikimedia.org/wikipedia/commons/e/ed/Utm-zones.jpg}.
DEFAULT: UTM zone at \link{mean} of \code{long}}
}
\value{
\code{sp::\link[sp]{CRS}} objects for one of: \cr
        - UTM projection with given zone\cr
        - Open street map (and google) mercator projection\cr
        - Latitude Longitude projection\cr
}
\description{
coordinate reference system (CRS) Object for several proj4 character strings.
\code{posm} and \code{pll} are taken directly from
\code{OpenStreetMap::\link[OpenStreetMap]{osm}} and
\code{\link[OpenStreetMap]{longlat}}.
}
\examples{
posm()
str(posm())
posm()@projargs
pll()
putm(5:14) # Germany
putm(zone=33) # Berlin

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Aug 2016
}
\seealso{
\code{\link{projectPoints}}, \code{\link{degree}}
}
\keyword{spatial}

