% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPoints.R
\name{projectPoints}
\alias{projectPoints}
\title{Project latlon points}
\usage{
projectPoints(lat, long, drop = FALSE, zone = 32,
  proj = paste0("+proj=utm +zone=", zone, "+ellps=WGS84 +datum=WGS84"))
}
\arguments{
\item{lat}{A vector of latitudes}

\item{long}{A vector of longitudes}

\item{drop}{Drop to lowest dimension? DEFAULT: FALSE (unlike projectMercator)}

\item{zone}{UTM zone, see e.g. \url{https://upload.wikimedia.org/wikipedia/commons/e/ed/Utm-zones.jpg}. DEFAULT: 32}

\item{proj}{Projection string to preject to. DEFAULT: UTM zone 32 (center Germany)}
}
\value{
data.frame with points in new projection
}
\description{
Project long lat points to e.g. UTM projection.
Basically copied from OpenStreetMap::projectMercator
}
\examples{
library("OpenStreetMap")
lat <- runif(100, 6, 12)
lon <- runif(100, 48, 58)
plot(lat,lon)
plot(projectMercator(lat,lon), main="Mercator")
plot(projectPoints(lat,lon), main="UTM32")

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2016
}
\seealso{
\code{\link{scaleBarOSM}}, \code{\link{help}}
}
\keyword{spatial}

