% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectPoints.R
\name{projectPoints}
\alias{projectPoints}
\title{Project latlon points}
\usage{
projectPoints(lat, long, drop = FALSE, zone = mean(long)\%/\%6 + 31,
  proj = paste0("+proj=utm +zone=", zone, "+ellps=WGS84 +datum=WGS84"),
  crs = CRS(proj))
}
\arguments{
\item{lat}{A vector of latitudes}

\item{long}{A vector of longitudes}

\item{drop}{Drop to lowest dimension? DEFAULT: FALSE (unlike projectMercator)}

\item{zone}{UTM zone, see e.g. \url{https://upload.wikimedia.org/wikipedia/commons/e/ed/Utm-zones.jpg}. DEFAULT: at mean of long}

\item{proj}{proj4 character string or CRS object to project to. DEFAULT: UTM projection at \code{zone}}

\item{crs}{Coordinate Reference System Object. If given, overrides \code{proj}. DEFAULT: CRS(proj)}
}
\value{
data.frame with points in new projection
}
\description{
Project long lat points to e.g. UTM projection.
Basically copied from OpenStreetMap::projectMercator
}
\examples{
library("OpenStreetMap")
lat <- runif(100, 6, 12)
lon <- runif(100, 48, 58)
plot(lat,lon)
plot(projectMercator(lat,lon), main="Mercator")
plot(projectPoints(lat,lon), main="UTM32")
stopifnot(all(projectPoints(lat,lon, crs=osm())==projectMercator(lat,lon)))

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2016
}
\seealso{
\code{\link{scaleBarOSM}}, \code{\link[OpenStreetMap]{projectMercator}}
}
\keyword{spatial}

