% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleBarOSM.R
\name{scaleBarOSM}
\alias{scaleBarOSM}
\title{scalebar for OSM plots}
\usage{
scaleBarOSM(x = 0.1, y = 0.9, length = 0.2, abslen = NA,
  unit = c("km", "m", "mi", "ft", "yd"), field = "rect", fill = NA,
  adj = c(0.5, 1.5), cex = par("cex"), col = 1, targs = NULL, lwd = 5,
  lend = 1, ...)
}
\arguments{
\item{x, y}{Relative position of left end of scalebar. DEFAULT: 0.1, 0.9}

\item{length}{Approximate relative length of bar. DEFAULT: 0.1}

\item{abslen}{Absolute length in \code{unit}s. DEFAULT: NA (computed internally from length)}

\item{unit}{Unit for computation and label. kilometer and meter as well as
miles, feet and yards are possible.
Note that the returned absolute length is in m. DEFAULT: "km"}

\item{field, fill, adj, cex, col}{Arguments passed to \code{\link[berryFunctions]{textField}}}

\item{targs}{List of further arguments passed to \code{\link[berryFunctions]{textField}}
like font, col, etc. DEFAULT: NULL}

\item{lwd, lend}{Line width and end style passed to \code{\link{segments}}. DEFAULT: 5,1}

\item{\dots}{Further arguments passed to \code{\link{segments}} like col}
}
\value{
invisible: coordinates of scalebar and label
}
\description{
Add a scalebar to default OpenStreetMap plots
}
\details{
This is a starting version and might be expanded heavily in customization options
}
\section{Warning}{
 Only works correctly for UTM maps, not for the Mercator projection!
                  Has not been extensively tested yet, please report bugs.
}
\examples{
library("OpenStreetMap")
d <- data.frame(long=c(12.95, 12.98, 13.22, 13.11), lat=c(52.40,52.52, 52.36, 52.45))
map <- pointsMap(d, utm=TRUE)
coord <- scaleBarOSM()  ; coord
scaleBarOSM(0.3, 0.05, unit="m")
scaleBarOSM(0.3, 0.5, unit="km", length=0.1)
scaleBarOSM(0.12, 0.28, abslen=20000)
scaleBarOSM(0.6, 0.8, unit="mi", col="red", targ=list(col="red"))

\dontrun{ ## Development stuff for mercator projection

map <- pointsMap(d)
scaleBarOSM(abslen=5000) # wrong for mercator projection. 5km is longer
SDMTools::Scalebar(x=1442638,y=6893871,distance=10000) # wrong as well
raster::scalebar(d=5000, xy=c(1442638,6893871)) # also wrong
raster::pointDistance ?

map2 <- pointsMap(d, map=map, utm=TRUE, pargs=list(pch=NA))
par(mfrow=c(1,2))
plot(map2, removeMargin=T)
scaleBarOSM(x=0.33, y=0.48, abslen=5000)
plot(map, removeMargin=T)
mapmisc::scaleBar(map$tiles[[1]]$projection, seg.len=10, pos="center", bg="transparent")
# Well that's fairly good in some circumstances, but weird for others
}

}
\author{
Berry Boessenkool, \email{berry-b@gmx.de}, Jun 2016
}
\seealso{
\code{\link{projectPoints}}
}
\keyword{aplot}
\keyword{spatial}

